zephyr: Improve and rename img_mgmt_find_best_area_id
The img_mgmt_find_best_area_id has been renamed to
img_mgmt_get_unused_slot_area_id and now takes one integer parameter,
that is a slot number to check.
The parameter may be -1, in which case the function will check the
first two slots for availability to perform DFU.  In case when
positive integer is provided the function will address directly
selected slot, and will perform the same tests, as in auto-select,
for first two slots, but in case of slot numbers above 1, it will
just check if the slot has been defined.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>